[#53] Clean-up and Refactor Automation Spec Base Class for OpenPF2 Stds.
This officially adopts a fork of the automation spec code from the
engine as part of OpenPF2, including the `BeforeAll` modifications. This
gives us greater control over how these tests are documented, defined,
and executed, at the cost of having to maintain this on our own should
the original code in the engine change.